+++ ICD.Connect.Protocol.Tests/SerialBuffers/JsonSerialBufferTest.cs
@@ -43,6 +43,26 @@ namespace ICD.Connect.Protocol.Tests.SerialBuffers
 	}
 }";
 
+		private const string EXAMPLE_JSON_PARTIAL_A = @"ghfghfghfh}
+{
+    ""m"": ""So"",
+	""d"": 98
+}
+{";
+
+		private const string EXAMPLE_JSON_PARTIAL_B = @"	""m"": ""So"",
+	""d"": 99
+}
+{
+	""m"": ""S"",
+	""d"": {
+		""T"": ""Digital"",";
+
+		private const string EXAMPLE_JSON_PARTIAL_C = @"		""No"": 73,
+		""V"": true
+	}
+}jhgdjh}";
+
 		[Test]
 		public void CompletedStringEventTest()
 		{
@@ -65,6 +85,21 @@ namespace ICD.Connect.Protocol.Tests.SerialBuffers
 			buffer.OnCompletedSerial += (sender, e) => receivedEvents.Add(e);
 
 			buffer.Enqueue(EXAMPLE_JSON_MULTIPLE);
+			Assert.AreEqual(3, receivedEvents.Count);
+		}
+
+		[Test]
+		public void CompletedStringEventPartialTest()
+		{
+			List<StringEventArgs> receivedEvents = new List<StringEventArgs>();
+			JsonSerialBuffer buffer = new JsonSerialBuffer();
+
+			buffer.OnCompletedSerial += (sender, e) => receivedEvents.Add(e);
+
+			buffer.Enqueue(EXAMPLE_JSON_PARTIAL_A);
+			buffer.Enqueue(EXAMPLE_JSON_PARTIAL_B);
+			buffer.Enqueue(EXAMPLE_JSON_PARTIAL_C);
+
 			Assert.AreEqual(3, receivedEvents.Count);
 		}
 	}