+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Added program number to port debug lines
 
+### Changed
+ - Performance improvements for JsonSerialBuffer
+
 ## [8.0.0] - 2019-01-10
 ### Added
  - Added interfaces for port configuration